- Refactor of frame buffer - improve matlab type handling

+++ matlab/show_mesh.m
@@ -1,3 +1,3 @@
 function show_mesh(satellite, triangle_visbility, velocity__m_per_s)
-    MexGateway("Visualization.show_mesh", satellite.handle_, double(triangle_visbility), velocity__m_per_s);
+    MexGateway("Visualization.show_mesh", int32(satellite.handle_), double(triangle_visbility), velocity__m_per_s);
 end
